Move PublicMode into the public module
Its only two consumers are public-module containers, so it gets the same
treatment as the SDK modes rather than surviving as a hollowed-out mode in
shared-tier click-actions.
