+++ frontend/src/metabase/public/PublicMode.ts
@@ -1,4 +1,4 @@
-import type { QueryClickActionsMode } from "../../types";
+import type { QueryClickActionsMode } from "metabase/visualizations/types";
 
 export const PublicMode: QueryClickActionsMode = {
   name: "public",
